+++ src/main/kotlin/org/jetbrains/changelog/tasks/PatchChangelogTask.kt
@@ -27,7 +27,8 @@ open class PatchChangelogTask : DefaultTask() {
         Changelog(extension).apply {
             get(extension.unreleasedTerm).let {
                 val header = it.getHeaderNode()
-                val versionHeader = extension.headerFormat().format(arrayOf(extension.version))
+                val arguments = extension.headerArguments.toTypedArray()
+                val versionHeader = "## " + extension.headerMessageFormat().format(arguments)
 
                 File(extension.path).writeText(content.run {
                     if (extension.keepUnreleasedSection) {